Persistent storage (Described in Bug 354734)

+++ rms/org.eclipse.ptp.rm.lml.core/src/org/eclipse/ptp/rm/lml/core/model/ITableColumnLayout.java
@@ -35,9 +35,7 @@ public interface ITableColumnLayout {
 	 * 
 	 * @return width of the table
 	 */
-	public int getWidth();
-
-	// TODO maybe also position???
+	public Double getWidth();
 
 	/**
 	 * @return if table column is active or not
